fix or silence compiler warnings about narrowing, unused variables, hiding, and unreachable code
(cherry picked from commit 454d11267dcbfe923fb576025de9d974a1ba8191)

+++ test/algorithm/for_test2.cpp
@@ -36,7 +36,7 @@ namespace
         //std::ostringstream out;
         int iii;
         int x = 0;
-        int size = v.size();
+        int size = static_cast<int>(v.size());
             for_(ref(iii) = 0, ref(iii) < size, ++ref(iii) )
             [ ref(x) += arg1[ref(iii)] ] (v);
             BOOST_TEST(x == 45);
